rename revert
